chore: migrated cleanup to typescript
