+++ package.json
@@ -24,10 +24,10 @@
     "postinstall": "husky install",
     "prepublishOnly": "pinst --disable",
     "postpublish": "pinst --enable",
-    "build": "node tools/cleanup dist && tsc",
+    "build": "tsnode tools/cleanup && tsc",
     "license-check": "node tools/license-check",
     "dep-check": "node tools/dep-check",
-    "clean": "node tools/cleanup",
+    "clean": "ts-node tools/cleanup",
     "package": "npm run build && npm pack",
     "test": "jest --no-cache --runInBand",
     "test:cov": "jest --coverage --no-cache --runInBand",
